changes to old frontend.
removed navbar and the other pages, just kept the homepage and functionality as it was during the MadData25 hackathon presentaiton
